Implement an HttpOutputWriter as part #26.
This work is not complete yet, but I'm putting it out there (maybe someone will look at it and give me advice). It seems that it probably make sense to always batch HTTP results. I can't imagine a backend expecting single results sent on individual HTTP request.

This shows the limit of the current design of batching. In this case, my first hunch was to implement HttpOutputWriter as non batching (in retrospective, it is obviously wrong). I now need to modify it into an BatchedOutputWriter. Changes are not that important, but still ... there should be a better way to compose these concerns. If anyone has an idea of a better design, I'd be happy for the feedback!